Rename `J2D.ts` to `J2Dv1.ts` and add comment explaining v1 vs v2
